Update
Enhancements:
#9 Admin accounts/Organisation accounts
#35 Items use icons from legend
#10 Finished user profile changes
Updated statistics on homepage
Fixed extra files/folder
Other minor improvements

Bug fixes:
#36 Pagination + searching display extra white space
#71 Image upload is hardcoded to item 3
#70 Uploading image from camera error
#68 Create item modal org address still selected after org removed from
user
#72 Fails to redirect after creating item
#73 Homepage
Other minor bugs

+++ website/php/accountDelete.php
@@ -3,27 +3,40 @@
 	session_start();
 
 	if(isset($_SESSION['userID'])) {
-		$stmt = $db -> prepare ('DELETE FROM item WHERE FKclient=?');
-		$stmt -> execute (array($_SESSION['userID']));
+		if($_SESSION['userID'] === $_GET['id'] || $_SESSION['accountType'] === "3"){
+			$stmt = $db -> prepare ('DELETE FROM item WHERE FKclient=?');
+			$stmt -> execute (array($_GET['id']));
 
-		$stmt = $db -> prepare ('DELETE FROM notification WHERE FKclient=?');
-		$stmt -> execute (array($_SESSION['userID']));
+			$stmt = $db -> prepare ('DELETE FROM notification WHERE FKclient=?');
+			$stmt -> execute (array($_GET['id']));
 
-		$stmt = $db -> prepare ('DELETE FROM client WHERE clientID=?');
-		$stmt -> execute (array($_SESSION['userID']));
-		
-		if($stmt -> rowCount() > 0) {
-			session_destroy();
-			echo "success";	
+			$stmt = $db -> prepare ('DELETE FROM client WHERE clientID=?');
+			$stmt -> execute (array($_GET['id']));
+			
+			if($stmt -> rowCount() > 0) {
+				if ($_SESSION['accountType'] != "3"){
+					session_destroy();
+				}
+				echo "success";	
+			} else {
+				//Notify the user of the failure
+				echo '<br><div class="alert alert-danger alert-dismissible fade in" role="alert">
+	               <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+	                  <span aria-hidden="true">×</span>
+	               </button>
+	                  <p>Error: There was an error deleting your account</p>
+	                  <p><button type="button" class="btn btn-danger" data-dismiss="alert">Dismiss</button></p>
+	            </div>';
+			}
 		} else {
-			//Notify the user of the failure
+			//Notify the user they do not have permission to delete the account
 			echo '<br><div class="alert alert-danger alert-dismissible fade in" role="alert">
-               <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                  <span aria-hidden="true">×</span>
-               </button>
-                  <p>Error: There was an error deleting your account</p>
-                  <p><button type="button" class="btn btn-danger" data-dismiss="alert">Dismiss</button></p>
-            </div>';
+		       <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+		          <span aria-hidden="true">×</span>
+		       </button>
+		          <p>Error: You do not have permission to delete this account</p>
+		          <p><button type="button" class="btn btn-danger" data-dismiss="alert">Dismiss</button></p>
+		    </div>';
 		}
 	} else {
 		//Notify the user they are not logged in
@@ -31,7 +44,8 @@
            <button type="button" class="close" data-dismiss="alert" aria-label="Close">
               <span aria-hidden="true">×</span>
            </button>
-              <p>Error: You must be logged in to delete your account</p>
+              <p>Error: You are not logged in.</p>
+              <p>Please refresh the page then try again</p>
               <p><button type="button" class="btn btn-danger" data-dismiss="alert">Dismiss</button></p>
         </div>';
 	}
